% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdemog.R
\name{NOut}
\alias{NOut}
\title{Output population sizes at a particular time}
\usage{
NOut(time_pt, time.scale, demograph_out)
}
\arguments{
\item{time_pt}{A numeric value defining the time point for the migrations which will be plotted. \code{time_pt} should be in the scale defined by \code{time.scale}. For example, \code{time_pt} = 3 when \code{time.scale} = "log10year" corresponds to the migrations 10^3 years ago.}

\item{time.scale}{A keyword to define the time scale used in the plot. It can be "4Ne", "generation", "year", "kyear", and "log10year". When the \code{size.scale} = "topology", this parameter will be ignored.}

\item{demograph_out}{A list containing all demographic information, see \code{\link{PlotMS}}.}
}
\value{
A vector of all population sizes for the specified time.
}
\description{
Output population sizes at a particular time.
}
